Revert "[Chore] Add Larastan, improve CI, and add additional tests (#93)"
This reverts commit 64710b4d5300a2ca8f90c712716d0d432f843d3e.

+++ app/Actions/Insights/Crop/CreateCropEstimationMetric.php
@@ -78,7 +78,7 @@ class CreateCropEstimationMetric
 
     private function retrieveNewAmount(
         string $key,
-        float $estimatedYieldAmount,
+        int $estimatedYieldAmount,
         string $estimatedYieldDateUpperTag,
         string $estimatedYieldDateLowerTag,
         Crop $crop,
